fix(REMEDY-277): extract columns from reducer hash
